🚚 move components around to group under relevant pages
