+++ nativeApp/App.tsx
@@ -15,10 +15,10 @@ import { ColoursEnum } from './src/lib/ui/colours';
 import AdminRouter from './src/screens/admin_views/AdminRouter';
 import SettingsButton from './src/lib/ui/SettingsButton';
 import Settings from './src/screens/volunteer_views/Settings';
-import AdminAddProject from './src/screens/admin_views/AddProject';
-import AdminEditProject from './src/screens/admin_views/EditProject';
-import AdminEditTime from './src/screens/admin_views/EditTime';
-import AdminEditVolunteer from './src/screens/admin_views/EditVolunteer';
+import AdminAddProject from './src/screens/admin_views/Projects/AddProject';
+import AdminEditProject from './src/screens/admin_views/Projects/EditProject';
+import AdminEditTime from './src/screens/admin_views/AdminTime/EditTime';
+import AdminEditVolunteer from './src/screens/admin_views/Volunteers/EditVolunteer';
 
 
 const getSettingsButton = (navigation) => {
