+++ nativeApp/src/screens/admin_views/AdminRouter.tsx
@@ -1,14 +1,14 @@
 
-import React from 'react';
+import React, { ClassicComponent } from 'react';
 import { createAppContainer } from 'react-navigation';
 import { createBottomTabNavigator } from 'react-navigation-tabs';
 import { MaterialCommunityIcons, MaterialIcons } from '@expo/vector-icons';
 
 import { ColoursEnum } from '../../lib/ui/colours';
 import AdminHome from './AdminHome';
-import Projects from './Projects';
-import Volunteers from './Volunteers';
-import AdminTime from './AdminTime';
+import Projects from './Projects/ProjectsPage';
+import Volunteers from './Volunteers/VolunteersPage';
+import AdminTime from './AdminTime/AdminTimePage';
 import AddTime from './AddTime';
 
 const getIconName = (name: string) => {
