+++ nativeApp/src/screens/admin_views/AdminTime/AdminTimePage/index.tsx
@@ -1,9 +1,9 @@
 import React, { FC } from 'react';
 // import styled from 'styled-components/native';
 
-import { Heading } from '../../../lib/ui/typography';
-import TimeCard from '../../../lib/ui/TimeCard';
-import Page from '../../../lib/ui/Page';
+import { Heading } from '../../../../lib/ui/typography';
+import TimeCard from '../../../../lib/ui/TimeCard';
+import Page from '../../../../lib/ui/Page';
 
 /*
  * Types
