+++ nativeApp/src/screens/admin_views/Projects/AddProject/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 // import styled from 'styled-components/native';
-import Page from '../../../lib/ui/Page';
+import Page from '../../../../lib/ui/Page';
 
 /*
  * Types
@@ -12,14 +12,13 @@ type Props = {
  * Styles
  */
 
-
 /*
  * Component
  */
-const EditProject: FC<Props> = () => (
+const AddProject: FC<Props> = () => (
   <Page>
     {}
   </Page>
 );
 
-export default EditProject;
+export default AddProject;
