+++ nativeApp/src/screens/admin_views/Projects/EditProject/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 // import styled from 'styled-components/native';
-import Page from '../../../lib/ui/Page';
+import Page from '../../../../lib/ui/Page';
 
 /*
  * Types
@@ -12,13 +12,14 @@ type Props = {
  * Styles
  */
 
+
 /*
  * Component
  */
-const AddProject: FC<Props> = () => (
+const EditProject: FC<Props> = () => (
   <Page>
     {}
   </Page>
 );
 
-export default AddProject;
+export default EditProject;
