+++ nativeApp/src/screens/admin_views/Projects/ProjectsPage/index.tsx
@@ -1,9 +1,9 @@
 import React, { FC } from 'react';
 import styled from 'styled-components/native';
-import { Heading } from '../../../lib/ui/typography';
-import Tabs from '../../../lib/ui/Tabs';
+import { Heading } from '../../../../lib/ui/typography';
+import Tabs from '../../../../lib/ui/Tabs';
 import ProjectCard from './ProjectCard';
-import Page from '../../../lib/ui/Page';
+import Page from '../../../../lib/ui/Page';
 
 /*
  * Types
