+++ nativeApp/src/screens/admin_views/Volunteers/EditVolunteer/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 // import styled from 'styled-components/native';
-import Page from '../../../lib/ui/Page';
+import Page from '../../../../lib/ui/Page';
 
 /*
  * Types
